refactor: reduce code smells in script classes
