+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/outputs/scripts/ImportScript.java
@@ -1,37 +1,34 @@
-package fr.insee.kraftwerk.core.outputs;
-
-import java.util.LinkedHashMap;
-import java.util.Map;
-import java.util.Map.Entry;
+package fr.insee.kraftwerk.core.outputs.scripts;
 
 import fr.insee.kraftwerk.core.Constants;
 import fr.insee.kraftwerk.core.metadata.Variable;
 import fr.insee.kraftwerk.core.metadata.VariableType;
 import fr.insee.kraftwerk.core.metadata.VariablesMap;
-import fr.insee.vtl.model.Structured.DataStructure;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
+import fr.insee.vtl.model.Structured;
 
-/** POJO class to store information needed to write a script for a CSV table. */
-@AllArgsConstructor
-public class TableScriptInfo {
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
 
-	/** The name of the dataset in the destination language. */
-	@Getter
-	String tableName;
-	/** The CSV file name of the table. */
-	@Getter
-	String csvFileName;
-	/** The data structure (containing variable names and types) of the table. */
-	@Getter
-	DataStructure dataStructure;
+/**
+ * Class to generate import scripts for the csv output tables.
+ * Some methods that could be implemented later: R base script and Python with pandas script.
+ */
+public abstract class ImportScript {
 
-	/** The data structure (containing variable length) of the table. */
-	@Getter
-	Map<String, VariablesMap> metadataVariables;
+	final List<TableScriptInfo> tableScriptInfoList;
+	static final String END_LINE = "\n";
 
-	public Map<String, Variable> getAllLength(DataStructure dataStructure,
-			Map<String, VariablesMap> metadataVariables) {
+	/** @see TableScriptInfo */
+	protected ImportScript(List<TableScriptInfo> tableScriptInfoList) {
+		this.tableScriptInfoList = tableScriptInfoList;
+	}
+
+	public abstract String generateScript();
+
+	static Map<String, Variable> getAllLength(Structured.DataStructure dataStructure,
+											  Map<String, VariablesMap> metadataVariables) {
 		Map<String, Variable> result = new LinkedHashMap<>();
 		// datastructure : noms complets
 		// metadata : suffixe du nom
@@ -45,19 +42,20 @@ public class TableScriptInfo {
 
 				// We treat the identifiers
 				if (variablesMap.getIdentifierNames().contains(variableName) && !result.containsKey(variableName)) {
-					result.put(variableName, new Variable(variableName, variablesMap.getGroup(variableName),VariableType.STRING, "32"));					
+					result.put(variableName, new Variable(variableName, variablesMap.getGroup(variableName),VariableType.STRING, "32"));
 				}
 				if (variablesMap.getDistinctVariableNamesAndFullyQualifiedNames().contains(variableName)) {
-					
-						Variable variable = variablesMap.getVariable(getRootName(variableName));
-						variableName = getRootName(variableName);
-						String newLengthString = variable.getLength();
-						
-						// We already got the variable, so we check to see if the lengths are different -> take the maximum one then
+
+					variableName = getRootName(variableName);
+					Variable variable = variablesMap.getVariable(variableName);
+
+					String newLengthString = variable.getLength();
+
+					// We already got the variable, so we check to see if the lengths are different -> take the maximum one then
 					if (newLengthString == null && !variableName.toUpperCase().contains(Constants.FILTER_RESULT_PREFIX)) {
 						if (result.containsKey(variableName)) {
 							result.replace(variableName, new Variable(variableName,
-									 result.get(variableName).getGroup(), VariableType.STRING, "255"));
+									result.get(variableName).getGroup(), VariableType.STRING, "255"));
 						} else {
 							result.put(variableName, new Variable(variableName,
 									variablesMap.getGroup(Constants.ROOT_GROUP_NAME), VariableType.STRING, "255"));
@@ -87,7 +85,6 @@ public class TableScriptInfo {
 							}
 
 						}
-
 					}
 				} else {
 					if (!result.containsKey(variableName)) {
@@ -102,7 +99,7 @@ public class TableScriptInfo {
 	}
 
 	/** Return the variable name without the group in the prefixes. */
-	public String getRootName(String name) {
+	static String getRootName(String name) {
 		return name.substring(name.lastIndexOf('.') + 1);
 	}
