+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/outputs/scripts/RImportScript.java
@@ -1,0 +1,16 @@
+package fr.insee.kraftwerk.core.outputs.scripts;
+
+import java.util.List;
+
+public class RImportScript extends ImportScript {
+
+    protected RImportScript(List<TableScriptInfo> tableScriptInfoList) {
+        super(tableScriptInfoList);
+    }
+
+    @Override
+    public String generateScript() {
+        return "Not implemented.";
+    }
+
+}
